WIP - Initial Drydock Chart
This will serve as the basis for the Drydock helm chart. As
functionality is added, the chart will continue to evolve. One
point we need to discuss is how to handle the image moving forward.

The image is currently in a personal dockerhub repository, but
I'll open a WIP pull request to the Dockerfiles repository and
update it as functionality is added to Drydock.