Fix Airflow web_server Link
Update the template so that the web_server
link does not appear with a "hash" tag
in-front-of it in the actual config file.

The problem arises after a major change to
the Shipyard Helm chart.  This Patch Set is
needed to correct that behavior.

We will make use of endpoints pattern from
Helm Toolkit to correct the bug.

Change-Id: I6d27643adfef666df579bcfcf7de166107292c34