Refactor Airflow Helm Chart
There is a need to align the Airflow Helm Chart with the recent
changes in OSH and to make use of Helm Toolkit instead.

This P.S. is meant to update the Airflow Helm Chart which was
created a while back. There should not be any changes to its
existing functionalities and things should still work as it is.

Change-Id: Ic9ab81df4ef7067acacd8b3499d95b4054acdfd0