Merge Shipyard and Airflow Helm Chart
It is clear that there is a tight coupling between Airflow and Shipyard.
For instance, we will need the dags, operators as well as shipyard.conf
for Airflow to function properly.

In order to avoid having to perform multiple workarounds to get the
workflow to work properly in Airflow, it has been decided that we will
merge the 2 Charts.

Note that a few changes have been introduced in the combined Chart

1) We have created a sidecar container with shipyard image so that all
   the dags and operators in the Master Branch can be copied to the
   Airflow Containers. This will be used for Production Environment.
   Note that we will continue to use the local directories on the local
   host machine for Dev Environment.

2) Introduce prod_environment option so that we can continue to access
   local host for making changes to the Dags and Operators by turning
   off that flag. Note that this is meant to aid development as it is
   time consuming to rebuild image with every change to the dag/operator

3) Include RabbitMQ as pre-req before starting Airflow Containers

4) Other fixes

Change-Id: I074fb4712bcf94aafeec62b3bc544321fa15f555