Bug Fix for Shipyard/Airflow Combined Chart
The below issues were identified when we conducted further testing.
This Patch Set is meant to fix these issues.

1) Shipyard.conf gets over-written and became read-only even though
   the environment was set to Lab
2) Shipyard container was starting prematurely and not adhering to
   the dependencies that were defined in values.yaml
3) Unable to execute 'concurrency check' due to incorrect admin user
   defined in values.yaml

Re-arranged dependencies in values.yaml to reflect the desired order
to start the containers.  Note that the shipyard API container should
only start after the Shipyard DB is initialized and populated.

Change-Id: I986de65ddcbc8055983475aa388a23f36b5ad8a4