Bug Fix for Shipyard/Airflow Helm Chart
The db-sync job was unable to read the airflow.cfg due to
incorrect mount path.

The home directory of airflow is '/usr/local/airflow/'.  This
is where the airflow.cfg should be placed so that airflow can
read and apply the configuration to populate the DB tables.

This P.S. is meant to correct the mount path

Change-Id: I836b828a49ab9338b54c8043a7f45f58e07ab996