Memcached Settings - Shipyard & DeckHand Chart
Update values.yaml and configmap-etc.yaml so that keystone
middleware will make use of memcached

This P.S. does the following:

1) Add memcached settings and configurations to shipyard.conf
   and deckhand.conf
2) Set 'enable_node_port' to false so that we can make use of
   over-ride to enable them if need be (we do not want these
   settings to be enabled by default).
3) Update shipyard.conf template - update 'keystone_auth' section
   and include 'logging' section
4) Point Airflow image to the latest image build in quay.io instead
   of the current private image registry

Note that there is a bug in KeystoneMiddleWare (Bug number 1713574
in launchpad) at the moment that will cause problems in Shipyard
when we try to trigger Shipyard API call with Memcached enabled.
This is due to compatibility issues between Python 2.x and 3.x. The
current work around will be to edit values.yaml and disable 'ENCRYPT'
in the 'memcache_security_strategy' parameter.  Note that this parameter
has been introduced in this Patch Set as part of memchaced requirements.

Tin will be fixing the KeystoneMiddleWare bug shortly and the issue will
get resolved after that.  Until then, we will need to apply the above
mentioned workaround for things to work properly in Shipyard.

Change-Id: I55cc680a842cbe0e68e3e42fcbe91ff483d7816b